# VideoGameApiWithKafka/Video Game API 1.9/openapi.yaml
openapi: 3.0.0
info:
  title: Video Game Exchange
  version: 1.0.0
servers:
- url: http://localhost:8080
paths:
...
  /games:
    get:
      operationId: listGames
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Game'
                type: array
          description: A paged array of Games
      summary: List All Games
      tags:
      - games
      x-accepts: application/json
      x-tags:
      - tag: games
    post:
      operationId: createGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
      summary: Create a Game
      tags:
      - games
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: games
  /games/{gameId}:
    delete:
      operationId: deleteGame
      parameters:
      - description: The id of the game to delete
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Delete one Game
      tags:
      - games
      x-accepts: application/json
      x-tags:
      - tag: games
    get:
      operationId: getOneGame
      parameters:
      - description: The id of the game to get
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
          description: One User Object
      summary: Get One Game
      tags:
      - games
      x-accepts: application/json
      x-tags:
      - tag: games
    patch:
      operationId: updateSomeGame
      parameters:
      - description: The id of the game to update
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Update Some Game Attributes
      tags:
      - games
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: games
    put:
      operationId: updateGame
      parameters:
      - description: The id of the game to update
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Update all Game Attributes
      tags:
      - games
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: games
...
components:
  schemas:
...
    Game:
      example:
        console: console
        numOwners: 6
        condition: condition
        pubishYear: pubishYear
        publisher: publisher
        id: 0
        title: title
      properties:
        id:
          format: int64
          type: integer
        title:
          type: string
        publisher:
          type: string
        pubishYear:
          type: string
        condition:
          type: string
        console:
          type: string
        numOwners:
          type: integer
      required:
      - condition
      - console
      - id
      - publishYear
      - publisher
      - title
      type: object
...